Fix 6 argument syscall inline assembly
It turns out the bug in the inline assembly was due to the `pushl` and
`popl` combo in the assembly. These instructions modify the ESP, and the
memory operand we get from GCC that we write to EBP is relative to the
stack pointer. So by modifying the stack pointer, the memory operand is
no longer valid.

The simple fix is to simply declare a temporary on the stack, and
provide that temporary to the inline assembly as a location to store the
EBP value at, rather then pushing it onto the stack. Declaring ESP as a
clobber might have worked, but would probably have bad implications on
the resulting code.